requirements.txt
Django
youtube-transcript-api
google-generativeai
django-cors-headers
gunicorn